feat: move 'tendermint_health_check' test utility function to autonomy.test_tools
This is a preparatory change in order to move test base classes into autonomy.test_tools
