+++ tests/test_agents/test_counter.py
@@ -31,9 +31,10 @@ import pytest
 import requests
 from aea.test_tools.test_cases import AEATestCaseMany
 
+from autonomy.constants import DEFAULT_REQUESTS_TIMEOUT
+
 from tests.conftest import ANY_ADDRESS
 from tests.fixture_helpers import UseTendermint
-from tests.helpers.constants import DEFAULT_REQUESTS_TIMEOUT
 from tests.helpers.tendermint_utils import (
     BaseTendermintTestClass,
     TendermintLocalNetworkBuilder,
